refactor(json): don't export rapidjson, make json methods private
Conflicts:
	src/json.cc
